// services/api/src/modules/tutor/enums/safety-category.enum.ts
export enum SafetyCategory {
  SAFE_EDUCATIONAL = 'SAFE_EDUCATIONAL',
  SELF_HARM = 'SELF_HARM',
  VIOLENCE = 'VIOLENCE',
  WEAPONS = 'WEAPONS',
  SEXUAL_CONTENT = 'SEXUAL_CONTENT',
  BULLYING = 'BULLYING',
  HATE = 'HATE',
  ILLEGAL_ACTIVITY = 'ILLEGAL_ACTIVITY',
  JAILBREAK = 'JAILBREAK',
  PERSONAL_DATA = 'PERSONAL_DATA',
  GENERAL_WELLBEING = 'GENERAL_WELLBEING',
}